Add Machine Whitelist handling
Behavior: If there are no registered machines, whitelist is disabled and any requests from any serial number is accepted. If there are at least one registered machine then it will look if the serial number is registered and if registered it will accept and continue the request, if not registered it will respond with a 403 Forbidden.

Also this is the start if the restructuring as seen by the different naming convention.

This also includes an updated .env example